requirements.txt
# Python dependencies for blog automation

# For parsing YAML frontmatter in .qmd files
PyYAML>=6.0.1
